role based access (RBA) for admin users

+++ src/main/java/com/example/springlearnings/controller/JournalController.java
@@ -4,15 +4,15 @@ import com.example.springlearnings.api.models.JournalPayload;
 import com.example.springlearnings.entity.Journal;
 import com.example.springlearnings.services.errorhandling.exceptions.UserDoesNotExistException;
 import com.example.springlearnings.services.interfaces.IJournalManagementService;
+import com.example.springlearnings.services.interfaces.IUserManagementService;
+import com.example.springlearnings.utils.ControllerUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 
 @RestController
@@ -20,18 +20,15 @@ import java.util.NoSuchElementException;
 public class JournalController {
     @Autowired
     private IJournalManagementService journalManagementService;
+    @Autowired
+    private IUserManagementService userManagementService;
+
     private final Logger LOGGER = LoggerFactory.getLogger(JournalController.class);
 
-    @GetMapping
-    public ResponseEntity<List<Journal>> getJournals() {
-        List<Journal> journalList = journalManagementService.getJournals();
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("X-total-count", String.valueOf(journalList.size()));
-        return new ResponseEntity<>(journalList, headers, HttpStatus.OK);
-    }
 
     @GetMapping(path = "/{id}")
     public ResponseEntity<Journal> getJournalById(@PathVariable String id) {
+        if (validJournalId(id)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         try {
             Journal journal = journalManagementService.getJournalById(id);
             return new ResponseEntity<>(journal, HttpStatus.OK);
@@ -43,20 +40,28 @@ public class JournalController {
     @PostMapping
     public ResponseEntity<String> createJournal(@RequestBody JournalPayload payload) throws UserDoesNotExistException {
         LOGGER.debug("JournalController::createJournal with payload: {}", payload);
-        String id = journalManagementService.createJournal(payload);
+        String id = journalManagementService.createJournal(payload, ControllerUtils.getUsernameFromSecurityContext());
         return new ResponseEntity<>(id, HttpStatus.CREATED);
     }
 
     @DeleteMapping(path = "/{id}")
     public ResponseEntity<String> removeJournal(@PathVariable String id) {
+        if (validJournalId(id)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         journalManagementService.deleteJournal(id);
         return new ResponseEntity<>(id, HttpStatus.NO_CONTENT);
     }
 
-    @PutMapping
-    public ResponseEntity<String> updateJournal(@RequestBody Journal payload) {
-        LOGGER.debug("JournalController::updateJournal with payload: {}", payload);
-        journalManagementService.updateJournal(payload.getId(), payload.getUsername(), payload.getContent());
-        return new ResponseEntity<>(payload.getId(), HttpStatus.CREATED);
+    @PutMapping(path = "/{id}")
+    public ResponseEntity<String> updateJournal(@PathVariable String id, @RequestBody String content) {
+        if (validJournalId(id)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        LOGGER.debug("JournalController::updateJournal with id: {}", id);
+        journalManagementService.updateJournal(id, content);
+        return new ResponseEntity<>(id, HttpStatus.CREATED);
+    }
+
+    private boolean validJournalId(String id) {
+        String username = ControllerUtils.getUsernameFromSecurityContext();
+        return userManagementService.getUserByUserName(username)
+                .getJournalList().stream().noneMatch(journal -> id.equalsIgnoreCase(journal.getId()));
     }
 }
